% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spurcore-package.R
\docType{package}
\name{spurcore-package}
\alias{spurcore}
\alias{spurcore-package}
\title{spurcore: Core Classes and Utilities for Spur Packages}
\description{
Core classes and utilities for spur packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://dexter.barrows.io/spurcore}
  \item \url{https://github.com/dbarrows/spurcore}
  \item Report bugs at \url{https://github.com/dbarrows/spurcore/issues}
}

}
\author{
\strong{Maintainer}: Dexter Barrows \email{dexter@barrows.io} (\href{https://orcid.org/0000-0001-8169-8938}{ORCID})

}
\keyword{internal}
