% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Volume class}
\usage{
volume(dims, h, seed = numeric())
}
\arguments{
\item{dims}{\code{\link{vector}} of the dimensions of the system in the order (x, y, z)}

\item{h}{the side length of each cubic voxel}

\item{seed}{optional \code{\link{vector}} containing the initial state value for each voxel}
}
\value{
An instance of the \code{\link{volume}} class
}
\description{
For setting initial conditions of inhomogeneous systems.
}
