% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_get}
\alias{volume_get}
\title{Get reaction system state}
\usage{
volume_get(volume, index)
}
\arguments{
\item{volume}{an instance of the \code{\link{volume}} class}

\item{index}{\code{\link{vector}} of the index (x, y, z) of the voxel to get the vector of species quantities}
}
\value{
\code{\link{vector}} of species quantities
}
\description{
Get reaction system state
}
