% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_h}
\alias{volume_h}
\title{Voxel side length}
\usage{
volume_h(volume)
}
\arguments{
\item{volume}{an instance of the \code{volume} class}
}
\value{
the side length of each voxel
}
\description{
Voxel side length
}
