% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_set}
\alias{volume_set}
\title{Set reaction system state}
\usage{
volume_set(volume, index, state)
}
\arguments{
\item{volume}{an instance of the \code{\link{volume}} class}

\item{index}{\code{\link{vector}} of the index (x, y, z) of the voxel to assign the state to}

\item{state}{\code{\link{vector}} of species quantities}
}
\description{
Set reaction system state
}
