% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_states}
\alias{volume_states}
\title{Volume states}
\usage{
volume_states(volume, species = NULL, ...)
}
\arguments{
\item{volume}{an instance of the \code{\link{volume}} class}

\item{species}{an optional \code{\link{vector}} of species names to be used in the output header}

\item{...}{additional options}
}
\value{
States of \code{\link{volume}} as a \code{\link[tibble:tibble]{tibble::tibble}}
}
\description{
Volume states
}
