% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Volume class for setting initial conditions of inhomogeneous systems}
\usage{
volume(dims, h, seed = numeric())
}
\arguments{
\item{dims}{the dimensions of the system in the order (x, y, z)}

\item{h}{the side length of each cubic voxel}

\item{seed}{optional vector containing the initial state value for each voxel in the order returned by \code{species(network)}}
}
\value{
an instance of the \code{volume} class
}
\description{
Volume class for setting initial conditions of inhomogeneous systems
}
