% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_dims}
\alias{volume_dims}
\title{Volume dimensions}
\usage{
volume_dims(volume)
}
\arguments{
\item{volume}{an instance of the \code{volume} class}
}
\value{
the dimensions of the system in the order (x, y, z)
}
\description{
Volume dimensions
}
