% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_get}
\alias{volume_get}
\title{Get reaction system state}
\usage{
volume_get(volume, index)
}
\arguments{
\item{volume}{an instance of the \code{volume} class}

\item{index}{the index (x, y, z) of the voxel to get the vector of species quantities}
}
\value{
vector of species quantities in the order returned by \code{species(network)}
}
\description{
Get reaction system state
}
