% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume_states}
\alias{volume_states}
\title{Volume states in table format}
\usage{
volume_states(volume, species = NULL, ...)
}
\arguments{
\item{volume}{an instance of the \code{volume} class}

\item{species}{an optional vector of species names to be used in the output header}

\item{...}{additional options}
}
\value{
the states of \code{volume} as a \code{tibble}
}
\description{
Volume states in table format
}
